+++ app/routes/_map/route.tsx
@@ -3,7 +3,6 @@ import type {
   Marker as TMarker,
   MarkerClusterGroup as TMarkerClusterGroup,
 } from "leaflet";
-import { clamp } from "lodash-es";
 import { EllipsisIcon, MapIcon } from "lucide-react";
 import React, { useImperativeHandle } from "react";
 import type { LoaderFunctionArgs } from "react-router";
@@ -48,6 +47,7 @@ import {
 import { useDralSafearea } from "./DralSafearea";
 import { createClusterIcon, createHighlightedClusterIcon } from "./Markers";
 import PinIcon from "./icons/PinIcon";
+import { handleLocation, handlePost, setHandled } from "./mapHelper";
 
 import tailwind_url from "~/app.css?url";
 
@@ -857,127 +857,3 @@ const PcsMarkers = ({
 };
 
 const PcsMarkersMemo = React.memo(PcsMarkers);
-
-const roundZoom = (z: number) => Math.round(z * 100) / 100;
-
-const setHandled = (
-  locationId: string | null,
-  postId: string | null,
-  locRef: React.RefObject<string | null>,
-  postRef: React.RefObject<string | null>
-) => {
-  locRef.current = locationId;
-  postRef.current = postId;
-};
-
-type Offsets = {
-  top: number;
-  left: number;
-  right: number;
-  bottom: number;
-};
-
-const handleLocation = async (
-  map: Map,
-  locationId: string,
-  marker: TMarker,
-  clusterGroup: TMarkerClusterGroup | null,
-  offsets: Offsets,
-  setHandledFn: () => void
-) => {
-  const coords = marker.getLatLng();
-  const center: [number, number] = [coords.lat, coords.lng];
-  const bounds = Offcenter.getBounds(map, offsets);
-  const cluster = clusterGroup?.getVisibleParent(marker);
-
-  if (cluster && cluster !== marker) {
-    const currentZoom = map.getZoom();
-    const parent = marker.__parent as MarkerCluster;
-    const clusterZoom = parent._zoom as number;
-
-    const targetZoom =
-      currentZoom >= 7 && parent.getChildCount() < 5
-        ? currentZoom
-        : clamp(clusterZoom + 1, Math.max(currentZoom, 8), map.getMaxZoom());
-
-    const newZoom = roundZoom(targetZoom);
-    const currentCenter = map.getCenter();
-
-    if (
-      newZoom !== roundZoom(currentZoom) ||
-      currentCenter.lat !== center[0] ||
-      currentCenter.lng !== center[1]
-    ) {
-      map.setView(
-        Offcenter.recenter(center, newZoom, offsets),
-        newZoom,
-        { animate: true }
-      );
-    }
-
-    await Promise.race([
-      new Promise((resolve) => map.once("moveend zoomend", resolve)),
-      new Promise((resolve) => setTimeout(resolve, 500)),
-    ]);
-
-    const finalCluster = clusterGroup?.getVisibleParent(marker);
-    if (finalCluster instanceof MarkerCluster) {
-      await new Promise((r) => setTimeout(r, 300));
-      finalCluster.spiderfy();
-    }
-
-    setHandledFn();
-  } else {
-    if (!bounds.contains(coords)) {
-      const currentZoom = roundZoom(map.getZoom());
-      map.setView(
-        Offcenter.recenter(center, currentZoom, offsets),
-        currentZoom,
-        { animate: true, duration: 0.7 }
-      );
-    }
-    setHandledFn();
-  }
-};
-
-export interface PostType {
-  id: string;
-  locations: {
-    id: string; // optional, but likely useful
-    data: {
-      coordinates: {
-        lat: number;
-        lng: number;
-      };
-    };
-  }[];
-}
-
-
-const handlePost = (
-  map: Map,
-  postId: string,
-  pcs_posts: PostType[],
-  sidebarWidth: number,
-  setHandledFn: () => void
-) => {
-  const post = pcs_posts.find((p) => p.id === postId);
-  if (!post || post.locations.length === 0) return;
-
-  const latlngs = post.locations.map((loc) => [
-    loc.data.coordinates.lat,
-    loc.data.coordinates.lng,
-  ]) as [number, number][];
-
-  const bounds = new LatLngBounds(latlngs);
-
-  console.log("🗺 fitBounds for post", postId, latlngs);
-
-  map.fitBounds(bounds, {
-    paddingTopLeft: [sidebarWidth + 80, 80],
-    paddingBottomRight: [80, 80],
-    animate: true,
-  });
-
-  setHandledFn();
-};
